+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - tauri: improve security #4826
 - improve fatal error dialog readability by removing color from deltachat-rpc-server errors
 - prevent dragging around of webxdc icon #4740
+- tauri: clear temp folder on exit #4839
 
 <a id="1_56_0"></a>